# source/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WNR-Audit API — Microsoft tenant security auditing
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants/{tenantId}/scans:
    parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listTenantScans
      tags: [scans]
      summary: List scans for a tenant
      responses:
        "200":
          description: Scans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scan"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: startTenantScan
      tags: [scans]
      summary: Start a new scan for a tenant
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartTenantScanBody"
      responses:
        "201":
          description: Scan started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "404":
          $ref: "#/components/responses/NotFound"
  /scans/{scanId}:
    parameters:
      - name: scanId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getScan
      tags: [scans]
      summary: Get scan details
      responses:
        "200":
          description: Scan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanDetails"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /tenants/{tenantId}/licenses:
    parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listTenantLicenses
      tags: [inquiries]
      summary: Listar licenças (subscribedSkus) do tenant via Microsoft Graph
      responses:
        "200":
          description: Licenças do tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantLicensesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
  /tenants/{tenantId}/inquiries:
    parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listTenantInquiries
      tags: [inquiries]
      summary: Listar consultas anteriores do tenant
      responses:
        "200":
          description: Lista de consultas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantInquiry"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createTenantInquiry
      tags: [inquiries]
      summary: Fazer uma nova consulta sobre serviços/configurações Microsoft 365
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInquiryBody"
      responses:
        "201":
          description: Consulta respondida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantInquiry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
  /tenants/{tenantId}/inquiries/{inquiryId}:
    parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
      - name: inquiryId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getTenantInquiry
      tags: [inquiries]
      summary: Detalhes de uma consulta
      responses:
        "200":
          description: Consulta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantInquiry"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    BadGateway:
      description: Bad gateway — upstream Microsoft API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Severity:
      type: string
      enum: [critical, high, medium, low, info]
    FindingStatus:
      type: string
      enum: [open, ignored, resolved]
...
    ScanStatus:
      type: string
      enum: [pending, running, completed, failed]
    ScanMode:
      type: string
      enum: [simple, advanced]
...
    TenantLicenseSku:
      type: object
      properties:
        skuId:
          type: string
        skuPartNumber:
          type: string
        displayName:
          type: string
          nullable: true
        prepaidUnits:
          type: integer
        consumedUnits:
          type: integer
        servicePlans:
          type: array
          items:
            type: object
            properties:
              servicePlanId:
                type: string
              servicePlanName:
                type: string
              provisioningStatus:
                type: string
            required: [servicePlanId, servicePlanName, provisioningStatus]
      required: [skuId, skuPartNumber, prepaidUnits, consumedUnits, servicePlans]
    TenantLicensesResponse:
      type: object
      properties:
        skus:
          type: array
          items:
            $ref: "#/components/schemas/TenantLicenseSku"
      required: [skus]
    TenantInquiry:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        question:
          type: string
        answer:
          type: string
        serviceDetected:
          type: string
          nullable: true
        licenseSummary:
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, question, answer, createdAt]
    CreateInquiryBody:
      type: object
      properties:
        question:
          type: string
          minLength: 3
      required: [question]
...
    Scan:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        status:
          $ref: "#/components/schemas/ScanStatus"
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        totalChecks:
          type: integer
        passedChecks:
          type: integer
        failedChecks:
          type: integer
        criticalCount:
          type: integer
        highCount:
          type: integer
        mediumCount:
          type: integer
        lowCount:
          type: integer
        scanMode:
          $ref: "#/components/schemas/ScanMode"
      required:
        - id
        - tenantId
        - status
        - startedAt
        - totalChecks
        - passedChecks
        - failedChecks
        - criticalCount
        - highCount
        - mediumCount
        - lowCount
        - scanMode
    StartTenantScanBody:
      type: object
      properties:
        mode:
          $ref: "#/components/schemas/ScanMode"
    ScanDetails:
      allOf:
        - $ref: "#/components/schemas/Scan"
        - type: object
          properties:
            tenantName:
              type: string
            findings:
              type: array
              items:
                $ref: "#/components/schemas/Finding"
            controls:
              type: array
              description: Catálogo completo dos controles avaliados, com status pass/fail.
              items:
                $ref: "#/components/schemas/ScanControl"
          required: [tenantName, findings, controls]
    ScanControl:
      type: object
      description: Item do catálogo CIS avaliado nesta varredura.
      properties:
        controlId:
          type: string
        title:
          type: string
        category:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        affectedResource:
          type: string
          nullable: true
        recommendation:
          type: string
          description: O que o tenant deve ter configurado para passar neste controle.
        status:
          type: string
          enum: [passed, failed]
        findingId:
          type: string
          nullable: true
          description: ID da vulnerabilidade aberta (apenas quando status=failed).
      required: [controlId, title, category, severity, recommendation, status]
    Finding:
      type: object
      properties:
        id:
          type: string
        scanId:
          type: string
        controlId:
          type: string
          description: CIS Benchmark control id (e.g. "1.1.1")
        title:
          type: string
        category:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        status:
          $ref: "#/components/schemas/FindingStatus"
        affectedResource:
          type: string
          nullable: true
        detectedAt:
          type: string
          format: date-time
      required:
        - id
        - scanId
        - controlId
        - title
        - category
        - severity
        - status
        - detectedAt
...